# openapi.yaml
openapi: 3.0.0
info:
  title: Ethereum Watcher API
  description: Create and Manage Logging Rules for the Ethereum blockchain
  version: 0.1.9
servers:
  - url: http://localhost:3000/
    description: dev server
paths:
  /rules/{id}:
    get:
      summary: Get a rule by id
      parameters:
        - name: id
          required: true
          schema:
            type: integer
          in: path
      responses:
        200:
          description: The Rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Set rule active
      parameters:
        - name: id
          required: true
          schema:
            type: integer
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - active
              additionalProperties: false
              properties:
                active:
                  type: boolean
                  nullable: false
      responses:
        200:
          description: The updated rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a rule by id
      parameters:
        - name: id
          required: true
          schema:
            type: integer
          in: path
      responses:
        200:
          description: Operation status
          content:
            application/json:
              schema:
                type: object
                required:
                  - deletedCount
                additionalProperties: false
                properties:
                  deletedCount:
                    type: integer
        default:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /rules:
    get:
      summary: Get a list of rules
      parameters:
        - name: active
          required: false
          schema:
            type: boolean
          in: query
      responses:
        default:
          description: The list of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
    post:
      summary: make a new watch rule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreationPayload'
      responses:
        default:
          description: the new rule
          content:
            application/json:
              schema:
                type: object

  /transactions:
    get:
      summary: Get a list of transactions
      parameters:
        - name: ruleId
          in: query
          schema:
            type: integer
        - name: hash
          in: query
          schema:
            type: string
        - name: blockHash
          in: query
          schema:
            type: string
        - name: blockNumber
          in: query
          schema:
            type: string
      responses:
        default:
          description: The list of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'

components:
  schemas:
    Error:
      type: object
      required:
        - status
      additionalProperties: false
      properties:
        status:
          type: integer
        message:
          type: string
    RuleCreationPayload:
      type: object
      additionalProperties: false
      required:
        - active
        - delay
        - criteria
      properties:
        active:
          type: boolean
          nullable: false
        delay:
          type: number
          nullable: false
        criteria:
          type: array
          minLength: 1
          nullable: false
          items:
            type: object
            nullable: false
      example:
        {
          active: true,
          delay: 0,
          criteria: [{ $and: [{ value: {$gt: 0} }, { gasPrice: {$gt: 0} }] }]
        }

    Rule:
      type: object
      additionalProperties: true
      required:
        - id
        - active
        - delay
        - criteria
      properties:
        id:
          type: integer
          nullable: false
        active:
          type: boolean
          nullable: false
        delay:
          type: integer
          nullable: false
        criteria:
          type: array
          items:
            type: object

    Transaction:
      type: object
      required:
        - ruleId
      additionalProperties: true
      properties:
        RuleId:
          type: integer
          nullable: false
        value:
          type: integer
        gasPrice:
          type: integer
        hash:
          type: string
        blockHash:
          type: string
        blockNumber:
          type: number